+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Feat: Improve search result rankings
 - Feat: Artist follower count
+- Feat: /me/followed-artists pagination
 - Fix: API improvements for integration with UI project
 
 ## [0.4.7](https://github.com/jaivalis/release-raccoon/compare/0.4.6...0.4.7)